+++ SpellCorrections.lua
@@ -23,6 +23,15 @@ local spellCorrectionsRetail = {
 	[277787] = 118, -- Tome of Polymorph: Direhorn
 	[277792] = 118, -- Tome of Polymorph: Bumblebee
 	[321395] = 118, -- Polymorph: Maw Rat
+	-- Arcane Torrent (Different SpellID per Class)
+	[28730] = 25046,
+	[50613] = 25046,
+	[69179] = 25046,
+	[80483] = 25046,
+	[129597] = 25046,
+	[155145] = 25046,
+	[202719] = 25046,
+	[232633] = 25046,
 	-- Druid Stampeding Roar
 	[77761] = 106898,
 	-- Ascendance
@@ -69,6 +78,12 @@ local spellCorrectionsRetail = {
 	[323673] = 375901,
 	-- Echoing Reprimand
 	[323547] = 385616,
+}
+
+local spellIconCorrections = {
+	-- These are icon ids!
+	-- Vengeance DH Meta
+	[135860] = 1247263
 }
 
 -- Wrath
@@ -161,9 +176,11 @@ local spellCorrectionsWrath = {
 	[20765] = 20707, -- Rank 5
 	[27239] = 20707, -- Rank 6
 	[47883] = 20707, -- Rank 7
+	-- Arcane Torrent (Different SpellID per Class / Power Type)
+	[28730] = 25046,
+	[50613] = 25046
 	--- General
-	-- Drums
-	[351355] = 35476, -- Greater Drums
+	--
 }
 
 -- Classic
@@ -245,6 +262,7 @@ local spellCorrectionsClassic = {
 }
 
 addon.spellCorrections = spellCorrectionsRetail
+addon.spellIconCorrections = spellIconCorrections
 
 --[===[@non-version-retail@
 spellCorrectionsRetail = nil